requirements.txt
django==5.1.4
django-ninja==1.3.0
django-ninja-extra==0.21.7
django-ninja-jwt==5.3.3
psycopg2-binary==2.9.10
gunicorn==23.0.0
whitenoise==6.8.2
